+++ ome2024-ngff-challenge/src/About.svelte
@@ -1,0 +1,13 @@
+<div>
+  <h2>About OME 2024 NGFF Challenge </h2>
+  <p>See the github link above and the
+    <a href="https://forum.image.sc/t/ome2024-ngff-challenge/97363">announcement on forum.image.sc</a>.
+  </p>
+</div>
+
+
+<style>
+  div {
+    margin: 15px;
+  }
+</style>
