+++ ome2024-ngff-challenge/src/Nav.svelte
@@ -1,19 +1,17 @@
-
 <script>
-  import omeMainNav from "/ome-main-nav.svg"
-
+  import omeMainNav from "/ome-main-nav.svg";
+  import githubLogo from "/github-mark-white.svg";
 </script>
 
 <nav>
-  <img style="height: 30px" alt="OME Logo" src={omeMainNav} />
-  <h2>2024 NGFF Challenge</h2>
+  <a class="logolink" href="/"><img alt="OME Logo" src={omeMainNav} /></a>
+  <a href="/"><h2>2024 NGFF Challenge</h2></a>
+  <div style="flex: 1"></div>
 
-  <ul>
-    <li>
-      <a href="https://github.com/ome/ome2024-ngff-challenge/" target="_blank">
-        About
-    </li>
-  </ul>
+  <a href="/about/"> About </a>
+  <a class="github" title="Open in GitHub" target="_blank" href="https://github.com/ome/ome2024-ngff-challenge/">
+    <img alt="GitHub Logo" src={githubLogo} />
+  </a>
 </nav>
 
 <style>
@@ -30,6 +28,7 @@
     align-items: center;
     box-sizing: border-box;
     flex: auto 0 0;
+    gap: 5px;
   }
 
   h2 {
@@ -37,25 +36,18 @@
     font-weight: 300;
     margin: 5px 15px;
   }
-
-  ul {
-    list-style: none;
-    color: white;
-    position: absolute;
-    right: 0;
-    top: 0;
-    margin: 10px 15px;
-  }
-
-  li {
-    float: right;
-    list-style-type: none;
-    margin: 0 10px;
-  }
   a {
     color: white;
     text-decoration: none;
-    line-height: 2.3;
+    display: block;
   }
-
+  .logolink, img {
+    height: 30px;
+  }
+  .github img, .github {
+    height: 20px;
+  }
+  .github {
+    margin: 0 10px;
+  }
 </style>
